Migrate fjs/media/json/tokenizer to .f.mjs
Splits JsonToken and the internal _ScanState/_ScanInput types into a
sibling types.ts. proof.f.ts stays TypeScript for now — it depends on
djs/serializer/module.f.ts, still unmigrated — and switches to
importing runtime values from module.f.mjs and types from types.ts.

Updates the three dependents (fjs/media/json/module.f.ts,
fjs/media/json/parser/module.f.ts, fjs/media/json/parser/proof.f.ts)
and stale doc references across fjs/djs/todo, fjs/js/todo, and
fjs/media/json/todo.
