+++ fjs/djs/todo/157.md
@@ -133,7 +133,7 @@ above and may land separately.
 > deleted and replaced by a grammar-based tokenizer. Re-verify current duplication
 > before acting on this section.
 
-`fjs/media/json/tokenizer/module.f.ts` and the old DJS tokenizer both wrapped the
+`fjs/media/json/tokenizer/module.f.mjs` and the old DJS tokenizer both wrapped the
 shared JS tokenizer and folded a leading `-` into the following numeric token.
 The current DJS tokenizer retains similar mapping/state logic, but the names and
 line numbers changed. Any extraction here must first re-measure the current code.
