+++ fjs/media/json/module.f.ts
@@ -17,7 +17,7 @@ import { concat } from '../../types/string/module.f.mjs'
 import { stringToList } from '../../text/utf16/module.f.mjs'
 import type { Result } from '../../types/result/types.ts'
 import { parse as parseTokens } from './parser/module.f.ts'
-import { tokenize } from './tokenizer/module.f.ts'
+import { tokenize } from './tokenizer/module.f.mjs'
 import { at, definedEntries } from '../../types/object/module.f.mjs'
 import type { Entry as ObjectEntry } from '../../types/object/types.ts'
 import { compose, fn } from '../../types/function/module.f.mjs'
