+++ fjs/media/json/parser/module.f.ts
@@ -8,7 +8,7 @@ import { error, ok } from '../../../types/result/module.f.mjs'
 import type { List } from '../../../types/list/types.ts'
 import { fold, next, toArray, concat } from '../../../types/list/module.f.mjs'
 import type { Fold } from '../../../types/function/operator/types.ts'
-import { type JsonToken } from '../tokenizer/module.f.ts'
+import type { JsonToken } from '../tokenizer/types.ts'
 import { setReplace } from '../../../types/ordered_map/module.f.mjs'
 import type { OrderedMap } from '../../../types/ordered_map/types.ts'
 import { type Unknown } from '../module.f.ts'
